DocumentCard component added
