+++ Swen3.Frontend/app/dashboard/page.tsx
@@ -5,6 +5,7 @@ import styles from './page.module.css';
 import { buildApiUrl } from '../utils/utils';
 import { CONSTANTS } from '../utils/constants';
 import { Document } from '../models/Document';
+import DocumentCard from '../components/DocumentCard';
 
 export default function Dashboard() {
   const [documents, setDocuments] = useState<Document[]>([]);
@@ -113,56 +114,6 @@ export default function Dashboard() {
     };
   }, [previewUrl]);
 
-  // Delete document
-  const deleteDocument = async (id: string) => {
-    try {
-      const response = await fetch(`${documentsEndpoint}/${id}`, {
-        method: 'DELETE'
-      });
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`);
-      }
-
-      setDocuments(documents.filter(doc => doc.id !== id));
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to delete document');
-    }
-  };
-
-  const summarize = async (doc: Document) => {
-    try {
-      setSummarizing(true);
-      setSummaryContent(null);
-      setShowSummaryPopup(true);
-      setCurrentSummaryDocTitle(doc.title);
-      setError(null);
-
-      // Using the document ID in the request body
-      const response = await fetch(`${geminiEndpoint}/summarize`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(doc.metadata),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(errorText || `HTTP error! status: ${response.status}`);
-      }
-
-      const result = await response.json();
-      setSummaryContent(result.summaryText || 'No summary available.');
-
-    } catch (err) {
-      setSummaryContent(`Error: ${err instanceof Error ? err.message : 'Failed to fetch summary'}`);
-      setError(err instanceof Error ? err.message : 'Failed to get document summary');
-    } finally {
-      setSummarizing(false);
-    }
-  }
-
   const closeSummaryPopup = () => {
     setShowSummaryPopup(false);
     setSummaryContent(null);
@@ -208,28 +159,6 @@ export default function Dashboard() {
   };
 
   const fileSizeUsage = selectedFile ? (selectedFile.size / CONSTANTS.MAX_FILE_BYTES) * 100 : 0;
-
-  const handleDownload = async (id: string) => {
-    try {
-      const doc = documents.find(d => d.id === id);
-      const downloadUrl = buildApiUrl(`/api/Documents/${id}/content`);
-
-      const response = await fetch(downloadUrl);
-      if (!response.ok) throw new Error('Download failed');
-
-      const blob = await response.blob();
-      const url = window.URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = doc?.fileName || 'document.pdf';
-      document.body.appendChild(a);
-      a.click();
-      window.URL.revokeObjectURL(url);
-      document.body.removeChild(a);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to download document');
-    }
-  };
 
   const handlePreview = (id: string) => {
     console.log('Opening preview for document:', id);
@@ -400,56 +329,7 @@ export default function Dashboard() {
         ) : (
           <div className={styles.documentsGrid}>
             {documents.map((doc) => (
-              <div key={doc.id} className={styles.docCard}>
-                <h3 className={styles.docTitle}>{doc.title}</h3>
-                <div className={styles.docMeta}>
-                  <div>📁 {doc.fileName}</div>
-                  <div>📅 {new Date(doc.uploadedAt).toLocaleDateString()}</div>
-                  <div>🏷️ {doc.mimeType}</div>
-                  <div className={styles.sizePill}>
-                    {(doc.size / (1024 * 1024)).toFixed(2)} MB
-                  </div>
-                </div>
-
-                {/* Metadata / OCR Text Section */}
-                <div className={styles.metadataSection}>
-                  <div className={styles.metadataLabel}>
-                    📝 OCR Text
-                  </div>
-                  {doc.metadata && doc.metadata.trim() !== '' ? (
-                    <div className={styles.metadataContent}>
-                      {doc.metadata}
-                    </div>
-                  ) : (
-                    <div className={styles.metadataEmpty}>
-                      Kein OCR-Text verfügbar
-                    </div>
-                  )}
-                </div>
-
-                {/* Document Actions */}
-                <div className={styles.docActions}>
-                  <button
-                    onClick={() => summarize(doc)}
-                    className={styles.btnSummarize}
-                    disabled={summarizing}
-                  >
-                    📝 Summarize
-                  </button>
-                  <button
-                    onClick={() => handleDownload(doc.id)}
-                    className={styles.btnDownload}
-                  >
-                    ⬇️ Download
-                  </button>
-                  <button
-                    onClick={() => deleteDocument(doc.id)}
-                    className={styles.btnDelete}
-                  >
-                    🗑️ Delete
-                  </button>
-                </div>
-              </div>
+              <DocumentCard {...doc} />
             ))}
           </div>
         )}
